requirements.txt
bcrypt==3.1.7
cffi==1.14.1
cryptography==3.0
fabric==2.5.0
invoke==1.4.1
paramiko==2.7.1
pycparser==2.20
PyNaCl==1.4.0
python-dotenv==0.14.0
six==1.15.0
